requirements.txt
dlt[bigquery]
pandas
requests
google-cloud-bigquery